🚚 Trekk klasser ut i egne filer for vedtaksstatistikk
