+++ src/main/kotlin/no/nav/tilleggsstonader/sak/behandling/domain/Behandling.kt
@@ -6,8 +6,6 @@ import no.nav.tilleggsstonader.sak.felles.domain.FagsakId
 import no.nav.tilleggsstonader.sak.infrastruktur.database.Sporbar
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.brukerfeilHvis
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.feilHvis
-import no.nav.tilleggsstonader.sak.statistikk.vedtak.BehandlingÅrsakDvh.MANUELT_OPPRETTET
-import no.nav.tilleggsstonader.sak.statistikk.vedtak.BehandlingÅrsakDvh.MANUELT_OPPRETTET_UTEN_BREV
 import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Column
 import org.springframework.data.relational.core.mapping.Embedded
