+++ src/main/kotlin/no/nav/tilleggsstonader/sak/statistikk/vedtak/domene/StønadstypeDvh.kt
@@ -1,0 +1,19 @@
+package no.nav.tilleggsstonader.sak.statistikk.vedtak.domene
+
+import no.nav.tilleggsstonader.kontrakter.felles.Stønadstype
+
+enum class StønadstypeDvh {
+    BARNETILSYN,
+    LÆREMIDLER,
+    ;
+
+    companion object {
+        fun fraDomene(stønadstype: Stønadstype): StønadstypeDvh {
+            return when (stønadstype) {
+                Stønadstype.BARNETILSYN -> BARNETILSYN
+                Stønadstype.LÆREMIDLER -> LÆREMIDLER
+                else -> error("Har ikke mapping for $stønadstype")
+            }
+        }
+    }
+}
